# src/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger - OpenAPI 3.0",
    "version": "1.0.11"
  },
  "tags": [
    {
      "name": "Products"
    },
    {
      "name": "Suppliers"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get list of all products",
        "responses": {
          "200": {
            "description": "A list of products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          },
          "400": {
            "description": "Invalid data"
          }
        }
      }
    },
    "/products/{productsId}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get single product by id",
        "responses": {
          "200": {
            "description": "Full information about this single product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductId"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          },
          "404": {
            "description": "No records found in the database matching your query."
          },
          "400": {
            "description": "Invalid data"
          }
        }
      }
    },
    "/suppliers": {
      "get": {
        "tags": ["Suppliers"],
        "summary": "Get list of all suppliers",
        "responses": {
          "200": {
            "description": "A list of suppliers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          },
          "400": {
            "description": "Invalid data"
          }
        }
      }
    },
    "/suppliers/{suppliersId}": {
      "get": {
        "tags": ["Suppliers"],
        "summary": "Get single supplier by id",
        "responses": {
          "200": {
            "description": "Full information about this single supplier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierId"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          },
          "404": {
            "description": "No records found in the database matching your query."
          },
          "400": {
            "description": "Invalid data"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "sqlLog": {
            "$ref": "#/components/schemas/SqlLog"
          },
          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer", "example": 1 },
                "ProductID": { "type": "integer", "example": 1 },
                "ProductName": { "type": "string", "example": "Chai" },
                "QuantityPerUnit": { "type": "string", "example": "10 boxes x 20 bags" },
                "UnitPrice": { "type": "string", "example": "18" },
                "UnitsInStock": { "type": "integer", "example": 39 },
                "UnitsOnOrder": { "type": "integer", "example": 0 }
              }
            }
          }
        }
      },
      "ProductId": {
        "type": "object",
        "properties": {
          "sqlLog": {
            "$ref": "#/components/schemas/SqlLog"
          },
          "product": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 1 },
              "ProductID": { "type": "integer", "example": 1 },
              "ProductName": { "type": "string", "example": "Chai" },
              "QuantityPerUnit": { "type": "string", "example": "10 boxes x 20 bags" },
              "UnitPrice": { "type": "string", "example": "18" },
              "UnitsInStock": { "type": "integer", "example": 39 },
              "SupplierID": { "type": "integer", "example": 1 },
              "ReorderLevel": { "type": "integer", "example": 10 },
              "Discontinued": { "type": "integer", "example": 0 },
              "UnitsOnOrder": { "type": "integer", "example": 0 },
              "Supplier": { "type": "string", "example": "Exotic Liquids" }
            }
          }
        }
      },
      "Supplier": {
        "type": "object",
        "properties": {
          "sqlLog": {
            "$ref": "#/components/schemas/SqlLog"
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer", "example": 1 },
                "SupplierID": { "type": "integer", "example": 1 },
                "CompanyName": { "type": "string", "example": "Exotic Liquids" },
                "ContactName": { "type": "string", "example": "Charlotte Cooper" },
                "ContactTitle": { "type": "string", "example": "Purchasing Manager" },
                "City": { "type": "string", "example": "London" },
                "Country": { "type": "string", "example": "UK" }
              }
            }
          }
        }
      },
      "SupplierID": {
        "type": "object",
        "properties": {
          "sqlLog": {
            "$ref": "#/components/schemas/SqlLog"
          },
          "supplier": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 2 },
              "SupplierID": { "type": "integer", "example": 2 },
              "CompanyName": { "type": "string", "example": "New Orleans Cajun Delights" },
              "ContactName": { "type": "string", "example": "Shelley Burke" },
              "ContactTitle": { "type": "string", "example": "Order Administrator" },
              "Address": { "type": "string", "example": "P.O. Box 78934" },
              "City": { "type": "string", "example": "New Orleans" },
              "Region": { "type": "string", "example": "North America" },
              "PostalCode": { "type": "string", "example": "70117" },
              "Country": { "type": "string", "example": "USA" },
              "Phone": { "type": "string", "example": "(100) 555-4822" }
            }
          }
        }
      },
      "SqlLog": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "querySqlLog": {
              "type": "string",
              "example": "SELECT \"id\", \"ProductID\", \"ProductName\", \"QuantityPerUnit\", \"UnitPrice\", \"UnitsInStock\", \"UnitsOnOrder\" FROM products;"
            },
            "startTime": { "type": "string", "example": "2023-06-22T13:18:57.665Z" },
            "finishTime": { "type": "string", "example": "2023-06-22T13:18:57.710Z" },
            "executionTimeToSecond": { "type": "integer", "example": 0.045 }
          }
        }
      }
    }
  }
}